renamed Game folder
