+++ game/author.rb
